MqGenerator字段映射修改，实体version字段: Integer -> Long

+++ sun8min-product/product-entity/src/main/java/com/sun8min/product/entity/ProductImage.java
@@ -67,7 +67,7 @@ public class ProductImage extends Model<ProductImage> {
      */
     @TableField("version")
     @Version
-    private Integer version;
+    private Long version;
 
     /**
      * 创建时间
