+++ sun8min-redpacket/redpacket-entity/src/main/java/com.sun8min.redpacket.entity/Redpacket.java
@@ -56,7 +56,7 @@ public class Redpacket extends Model<Redpacket> {
      */
     @TableField("version")
     @Version
-    private Integer version;
+    private Long version;
 
     /**
      * 创建时间
